+++ com.unity.cinemachine/CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Added CinemachineShotQualityEvaluator which is a standalone version of the evaluation code in Deoccluder.
 - Added StateDrivenCamera.CancelWait() method to cancel the current wait on a pending state change.
 - Added FlyAround sample scene showing a simple fly-around camera.
+- Added IgnoreTimeScale option to InputAxisControllerBase.
 
 ### Changed
 - SplineDolly and SplineCart: position on spline is preserved when units change.